requirements.txt
flask==2.0.1
opencv-python>=4.5.0
Pillow>=8.0.0
requests==2.26.0
python-dotenv==0.19.0
geopy==2.2.0
python-telegram-bot==13.7 